Skip the first 3 seconds of the dataset measurments
The measurements from the dataset include a rest period for about 5
seconds before they begin the movements. This seems to worsen the
training since each classification gets trained on the similar type of
resting data.

The researchers then trim from the front and back by some amount based
on the length of the "motion vector" to get a length of 13 seconds.
Inspecting the files, it seems like 3 seconds should be safe to cut
from most measurements.